feat: added Home component and refactored api folder
