--- src/api/index.js
@@ -1,3 +1,0 @@
-export { default as setupMockServer } from "./mock.server";
-export { default as useAxios } from "./useAxios";
-export { default as callMockServer } from "./server.request";
