+++ src/components/Home/Home.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import "./Home.css";
+
+function Home(props) {
+  return (
+    <div className="hero">
+      <div className="hero__img img--responsive"></div>
+      <div className="hero__container">
+        <div className="hero__heading h6 text--center text--bold">
+          Find what you love
+          
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Home;
