+++ src/components/index.js
@@ -1,16 +1,13 @@
 export { default as Product } from "./Product/Product";
 export { default as WishList } from "./WishList/WishList";
 export { default as Cart } from "./Cart/Cart";
+export { default as Home } from "./Home/Home";
 export { default as ProductListing } from "./Product/ProductListing";
 export { default as Toast } from "./Toast/Toast";
-export {
-  addItemToCart,
-  addOrRemoveFromWish,
-  removeFromCart,
-} from "./Product/ServerUpdate";
 export {
   ProductImage,
   ProductName,
   PrimaryButton,
   ProductContent,
 } from "./Product/ProductListing";
+export { handleToast } from "./Toast/Toast";
