+++ src/styles.css
@@ -1,4 +1,4 @@
-@import url("https://mind-ui-library.netlify.app/mind.ui.css");
+/* @import url("https://mind-ui-library.netlify.app/mind-ui.css"); */
 
 :root {
   --background-color: #ffffff;
